# Conflicts:
#	hedera-node/hedera-app/src/test/java/com/hedera/node/app/blocks/impl/streaming/BlockNodeCommunicationTestBase.java
#	hedera-node/hedera-app/src/test/java/com/hedera/node/app/blocks/impl/streaming/BlockNodeConnectionManagerTest.java
#	hedera-node/hedera-app/src/test/java/com/hedera/node/app/blocks/impl/streaming/BlockNodeConnectionTest.java
#	hedera-node/hedera-app/src/test/java/com/hedera/node/app/blocks/impl/streaming/BlockStreamStateManagerTest.java

+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/blocks/impl/streaming/BlockNodeCommunicationTestBase.java
@@ -81,7 +81,7 @@ public abstract class BlockNodeCommunicationTestBase {
         return createConfigProvider(BATCH_SIZE);
     }
 
-    protected ConfigProvider createConfigProvider(final long batchSize) {
+    protected ConfigProvider createConfigProvider(final int batchSize) {
         final var configPath = Objects.requireNonNull(
                         BlockNodeCommunicationTestBase.class.getClassLoader().getResource("bootstrap/"))
                 .getPath();
